Add metadata titles for different pages
