+++ client/package.json
@@ -47,6 +47,7 @@
     "eslint": "8.42.0",
     "eslint-config-next": "13.4.5",
     "express": "^4.17.2",
+    "lodash": "^4.17.21",
     "lucide-react": "^0.252.0",
     "mapbox-gl": "npm:empty-npm-package@1.0.0",
     "maplibre-gl": "3.3.0",
