+++ client/src/app/[section]/page.tsx
@@ -1,3 +1,6 @@
+import { startCase } from 'lodash';
+import type { Metadata } from 'next';
+
 import Hydrate from '@/lib/react-query/hydrate';
 
 import type { Section as SectionType } from '@/types/app';
@@ -6,8 +9,8 @@ import Providers from '@/app/page-providers';
 
 import Section from '@/containers/section';
 
+import { metadata as baseMetadata } from '../page';
 import { prefetchQueries } from '../prefetch';
-
 export default async function MapPage({
   params: { section },
 }: {
@@ -21,6 +24,17 @@ export default async function MapPage({
       </Providers>
     </Hydrate>
   );
+}
+
+export async function generateMetadata({
+  params: { section },
+}: {
+  params: { section: SectionType };
+}): Promise<Metadata> {
+  return {
+    ...baseMetadata,
+    title: `ORCaSa - ${startCase(section)}`,
+  };
 }
 
 export async function generateStaticParams() {
