+++ client/src/app/page.tsx
@@ -71,7 +71,21 @@ export default async function HomePage() {
           <LinkBox key={link.href} {...link} />
         ))}
       </div>
-      ;
     </div>
   );
 }
+
+// TODO: Update metadata
+export const metadata = {
+  title: 'ORCaSa',
+  description:
+    'Learn about ORCaSa, a Horizon Europe initiative that unites international stakeholders to develop techniques for capturing and storing carbon in soil. Join the effort to combat climate change today.',
+  keywords: 'ORCaSa, Horizon Europe, carbon capture, climate change',
+  robots: 'index, follow',
+  og: {
+    title: 'ORCaSa - Horizon Europe Initiative',
+    description:
+      'Learn about ORCaSa, a Horizon Europe initiative that unites international stakeholders to develop techniques for capturing and storing carbon in soil. Join the effort to combat climate change today.',
+    // image: 'https://example.com/orcasa-image.jpg',
+  },
+};
